\name{BNPdynIRT-package}
\alias{BNPdynIRT-package}
\alias{BNPdynIRT}
\docType{package}
\title{
\packageTitle{BNPdynIRT}
}
\description{
\packageDescription{BNPdynIRT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BNPdynIRT}
\packageIndices{BNPdynIRT}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BNPdynIRT}

Maintainer: \packageMaintainer{BNPdynIRT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
